fix syntax and move ngff_utils.py to ngff/utils.py
